amen_bank_backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.4.0
transformers==4.54.0
torch==2.7.1
nltk==3.9.1
requests==2.32.4
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3